+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @revstackhq/cli
 
+## 0.0.0-dev-20260227103607
+
+### Minor Changes
+
+- feat(cli, core): implement overage pricing configuration in PriceDef and business templates
+
+### Patch Changes
+
+- Updated dependencies
+  - @revstackhq/core@0.0.0-dev-20260227103607
+
 ## 0.0.0-dev-20260227093823
 
 ### Patch Changes