+++ packages/node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @revstackhq/node
 
+## 0.0.0-dev-20260227103607
+
+### Patch Changes
+
+- Updated dependencies
+  - @revstackhq/core@0.0.0-dev-20260227103607
+
 ## 0.0.0-dev-20260227093823
 
 ### Patch Changes